fix: inclui uso do utils nos exemplos de pipelines
