+++ pipeline/imdb.py
@@ -3,8 +3,9 @@ from learning_orchestra_client.function.python import FunctionPython
 from learning_orchestra_client.model.scikitlearn import ModelScikitLearn
 from learning_orchestra_client.train.scikitlearn import TrainScikitLearn
 from learning_orchestra_client.predict.scikitlearn import PredictScikitLearn
+from learning_orchestra_client.utils.file_caster import python_file_caster
 
-CLUSTER_IP = "http://34.123.167.241"
+CLUSTER_IP = "http://35.238.145.237"
 
 dataset_csv = DatasetCsv(CLUSTER_IP)
 
@@ -16,17 +17,7 @@ dataset_csv.insert_dataset_sync(
 
 function_python = FunctionPython(CLUSTER_IP)
 
-explore_dataset = '''
-pos=data[data["label"]=="1"]
-neg=data[data["label"]=="0"]
-
-total_rows = len(pos) + len(neg)
-
-print("Positive = " + str(len(pos) / total_rows))
-print("Negative = " + str(len(neg) / total_rows))
-
-response = None
-'''
+explore_dataset = python_file_caster('imdb/explore_dataset.py')
 
 function_python.run_function_sync(
     name="sentiment_analysis_exploring",
@@ -39,58 +30,7 @@ print(function_python.search_execution_content(
     skip=1,
     pretty_response=True))
 
-dataset_preprocessing = '''
-import re;
-
-
-def preprocessor(text):
-    global re
-    text = re.sub("<[^>]*>", "", text)
-    emojis = re.findall("(?::|;|=)(?:-)?(?:\)|\(|D|P)", text)
-    text = re.sub("[\W]+", " ", text.lower()) + \
-           " ".join(emojis).replace("-", "")
-    return text
-
-
-data["text"] = data["text"].apply(preprocessor)
-
-from nltk.stem.porter import PorterStemmer
-
-porter = PorterStemmer()
-
-
-def tokenizer_porter(text):
-    global porter
-    return [porter.stem(word) for word in text.split()]
-
-
-from sklearn.feature_extraction.text import TfidfVectorizer
-
-tfidf = TfidfVectorizer(strip_accents=None, 
-                        lowercase=False, 
-                        preprocessor=None,
-                        tokenizer=tokenizer_porter, 
-                        use_idf=True, 
-                        norm="l2",
-                        smooth_idf=True)
-
-y = data.label.values
-x = tfidf.fit_transform(data.text)
-
-from sklearn.model_selection import train_test_split
-
-X_train, X_test, y_train, y_test = train_test_split(x, y, 
-                                                    random_state=1,
-                                                    test_size=0.5,
-                                                    shuffle=False)
-        
-response = {
-    "X_train": X_train,
-    "X_test": X_test,
-    "y_train": y_train,
-    "y_test": y_test
-}
-'''
+dataset_preprocessing = python_file_caster('imdb/dataset_preprocessing.py')
 
 function_python.run_function_sync(
     name="sentiment_analysis_preprocessed",
@@ -141,13 +81,9 @@ predict_scikitlearn.create_prediction_sync(
 
 )
 
-logistic_regression_cv_accuracy = '''
-from sklearn import metrics
+logistic_regression_cv_accuracy = \
+    python_file_caster('imdb/logistic_regression_cv_accuracy.py')
 
-print("Accuracy: ",metrics.accuracy_score(y_test, y_pred))
-
-response = None
-'''
 function_python.run_function_sync(
     name="sentiment_analysis_logistic_regression_cv_accuracy",
     parameters={
