+++ pipeline/imdb/logistic_regression_cv_accuracy.py
@@ -1,0 +1,6 @@
+
+from sklearn import metrics
+
+print("Accuracy: ",metrics.accuracy_score(y_test, y_pred))
+
+response = None
