+++ pipeline/mnist.py
@@ -4,8 +4,8 @@ from learning_orchestra_client.model.tensorflow import ModelTensorflow
 from learning_orchestra_client.train.tensorflow import TrainTensorflow
 from learning_orchestra_client.predict.tensorflow import PredictTensorflow
 from learning_orchestra_client.evaluate.tensorflow import EvaluateTensorflow
-
-CLUSTER_IP = "http://35.224.50.116"
+from learning_orchestra_client.utils.file_caster import python_file_caster
+CLUSTER_IP = "http://35.238.145.237"
 
 dataset_generic = DatasetGeneric(CLUSTER_IP)
 dataset_generic.insert_dataset_async(
@@ -38,84 +38,8 @@ dataset_generic.wait("mnist_test_images")
 dataset_generic.wait("mnist_test_labels")
 
 function_python = FunctionPython(CLUSTER_IP)
-mnist_datasets_treatment = '''
-import numpy as np
-import struct as st
-import math
-
-training_filenames = {'images': mnist_train_images,
-                      'labels': mnist_train_labels}
-test_filenames = {'images': mnist_test_images, 
-                  'labels': mnist_test_labels}
-
-data_types = {
-    0x08: ('ubyte', 'B', 1),
-    0x09: ('byte', 'b', 1),
-    0x0B: ('>i2', 'h', 2),
-    0x0C: ('>i4', 'i', 4),
-    0x0D: ('>f4', 'f', 4),
-    0x0E: ('>f8', 'd', 8)}
-
-
-def treat_dataset(dataset: dict) -> tuple:
-    global np, st, math, data_types
-
-    for name in dataset.keys():
-        if name == 'images':
-            images_file = dataset[name]
-        if name == 'labels':
-            labels_file = dataset[name]
-
-    images_file.seek(0)
-    magic = st.unpack('>4B', images_file.read(4))
-    data_format = data_types[magic[2]][1]
-    data_size = data_types[magic[2]][2]
-
-    images_file.seek(4)
-
-    content_amount = st.unpack('>I', images_file.read(4))[0]
-    rows_amount = st.unpack('>I', images_file.read(4))[0]
-    columns_amount = st.unpack('>I', images_file.read(4))[0]
-
-    labels_file.seek(8)
-
-    labels_array = np.asarray(
-        st.unpack(
-            '>' + data_format * content_amount,
-            labels_file.read(content_amount * data_size))).reshape(
-        (content_amount, 1))
-
-    n_batch = 10000
-    n_iter = int(math.ceil(content_amount / n_batch))
-    n_bytes = n_batch * rows_amount * columns_amount * data_size
-    images_array = np.array([])
-
-    for i in range(0, n_iter):
-        temp_images_array = np.asarray(
-            st.unpack('>' + data_format * n_bytes,
-                      images_file.read(n_bytes))).reshape(
-            (n_batch, rows_amount, columns_amount))
-
-        if images_array.size == 0:
-            images_array = temp_images_array
-        else:
-            images_array = np.vstack((images_array, temp_images_array))
-
-        temp_images_array = np.array([])
-
-    return images_array, labels_array
-
-
-train_images, train_labels = treat_dataset(training_filenames)
-test_images, test_labels = treat_dataset(test_filenames)
-
-response = {
-    "train_images": train_images,
-    "train_labels": train_labels,
-    "test_images": test_images,
-    "test_labels": test_labels,
-}
-'''
+mnist_datasets_treatment = \
+    python_file_caster('mnist/mnist_datasets_treatment.py')
 
 function_python.run_function_async(
     name="mnist_datasets_treated",
@@ -128,23 +52,8 @@ function_python.run_function_async(
     code=mnist_datasets_treatment)
 function_python.wait("mnist_datasets_treated")
 
-mnist_datasets_normalization = '''
-test_images = test_images / 255
-train_images = train_images / 255
-
-print(train_images.shape)
-print(train_labels.shape)
-
-print(test_images.shape)
-print(test_labels.shape)
-
-response = {
-    "test_images": test_images,
-    "test_labels": test_labels,
-    "train_images": train_images,
-    "train_labels": train_labels
-}
-'''
+mnist_datasets_normalization = \
+    python_file_caster('mnist/mnist_datasets_normalization.py')
 
 function_python.run_function_async(
     name="mnist_datasets_normalized",
@@ -172,17 +81,7 @@ model_tensorflow.create_model_async(
 )
 model_tensorflow.wait("mnist_model")
 
-model_compilation = '''
-import tensorflow as tf
-
-model.compile(
-    optimizer=tf.keras.optimizers.Adam(0.001),
-    loss=tf.keras.losses.SparseCategoricalCrossentropy(),
-    metrics=[tf.keras.metrics.SparseCategoricalAccuracy()],
-)
-
-response = model
-'''
+model_compilation = python_file_caster('mnist/model_compilation.py')
 
 function_python.run_function_async(
     name="mnist_model_compiled",
@@ -233,10 +132,8 @@ evaluate_tensorflow.create_evaluate_async(
 predict_tensorflow.wait("mnist_model_predicted")
 evaluate_tensorflow.wait("mnist_model_evaluated")
 
-show_mnist_predict = '''
-print(mnist_predicted)
-response = None
-'''
+show_mnist_predict = python_file_caster('mnist/show_mnist_predict.py')
+
 function_python.run_function_async(
     name="mnist_model_predicted_print",
     parameters={
@@ -245,10 +142,8 @@ function_python.run_function_async(
     code=show_mnist_predict
 )
 
-show_mnist_evaluate = '''
-print(mnist_evaluated)
-response = None
-'''
+show_mnist_evaluate = python_file_caster('mnist/show_mnist_evaluate.py')
+
 function_python.run_function_async(
     name="mnist_model_evaluated_print",
     parameters={
