+++ pipeline/mnist/model_compilation.py
@@ -1,0 +1,9 @@
+import tensorflow as tf
+
+model.compile(
+    optimizer=tf.keras.optimizers.Adam(0.001),
+    loss=tf.keras.losses.SparseCategoricalCrossentropy(),
+    metrics=[tf.keras.metrics.SparseCategoricalAccuracy()],
+)
+
+response = model
