+++ pipeline/mnist/show_mnist_evaluate.py
@@ -1,0 +1,2 @@
+print(mnist_evaluated)
+response = None
