+++ pipeline/mnist/show_mnist_predict.py
@@ -1,0 +1,2 @@
+print(mnist_predicted)
+response = None
